kni: don't touch KNI device struct after freeing
Since the struct kni_dev is allocated as part of the netdev with
alloc_netdev(), when free_netdev() is called, this also frees the
struct kni_dev embedded in it.

This means that we cannot touch struct kni_dev after calling
free_netdev since that memory was already deallocated.

Separate unregistering the net_dev with unregister_netdev from freeing
the net_device and kni_dev structures into separate functions and
ensure that we do not touch anything in the kni_dev structure after
freeing it (ie don't call list_del(), etc...).

Signed-off-by: Dan Gora <dg@adax.com>