eal: choose random initial seed source at runtime
Instead of choosing to use getentropy or the rdseed instruction for
the random number generator entropy source using compilation flags,
determine the best source at run time.

This is accomplished by defining a weak symbol for getentropy, checking
that the compiler can generate the rdseed instruction even if the
compilation platform does not natively support it, and checking for
the availability of the rdseed instruction on the execution platform
using rte_cpu_get_flag_enabled.

If neither getentropy or rdseed is available, rte_get_timer_cycles
will be continue to be used as the entropy source.

Signed-off-by: Dan Gora <dg@adax.com>