[docker] Update to set the machine type and LTO from docker build cmd line.
Set the CONFIG_RTE_MACHINE variable based on the "machine_type"
build time argument.  This allows DPDK to be built for different
target architectures with different instruction sets.

Set CONFIG_RTE_ENABLE_LTO from the enable_lto variable.  We don't
need to modify the .config file, just set it on the make command line.

Labels are added to the image so that we can identify which docker
image is which.